Dash line & distance & transition gap
1.  When hover the nodes, the travel history is now showing as dash
line. The longer the appearance of line part is, the more recent the
path is.
2. The travel distance of one node is calculated based on the city's geo
location it travels. When hover a node, it will be shown on distance
entry.
3.  Now the time between each transition can be set in the query box.
The default value is 1500, the valid range is [100, 5000]. I recommend
the transition gap is better to be set no less than 500. Otherwise, each
transition process may not have enough time to finish.